# src/Presentation.Web.Server/wwwroot/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GettingsStarted API",
    "version": "v1.0"
  },
  "paths": {
    "/api/coremodule/customers/{id}": {
      "get": {
        "tags": [
          "CoreModule.Customers"
        ],
        "description": "Gets a customer by its unique identifier.",
        "operationId": "CoreModule.Customers.GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CoreModule.Customers"
        ],
        "description": "Updates the details of an existing customer identified by its unique identifier.",
        "operationId": "CoreModule.Customers.Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CoreModule.Customers"
        ],
        "description": "Deletes an existing customer identified by its unique identifier.",
        "operationId": "CoreModule.Customers.Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/coremodule/customers": {
      "get": {
        "tags": [
          "CoreModule.Customers"
        ],
        "description": "Gets all customers matching the specified filter criteria.",
        "operationId": "CoreModule.Customers.GetAll",
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/FilterModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerModel"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CoreModule.Customers"
        ],
        "description": "Creates a new customer with the provided details.",
        "operationId": "CoreModule.Customers.Create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/coremodule/customers/search": {
      "post": {
        "tags": [
          "CoreModule.Customers"
        ],
        "description": "Searches for customers matching the specified filter criteria.",
        "operationId": "CoreModule.Customers.Search",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerModel"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/coremodule/customers/{id}/status": {
      "put": {
        "tags": [
          "CoreModule.Customers"
        ],
        "description": "Updates the status of an existing customer identified by its unique identifier.",
        "operationId": "CoreModule.Customers.UpdateStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerUpdateStatusRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/coremodule/customers/exception": {
      "delete": {
        "tags": [
          "CoreModule.Customers"
        ],
        "description": "Endpoint that always throws an exception for testing purposes.",
        "operationId": "CoreModule.Customers.Delete2",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CompositeSpecification": {
        "type": "object",
        "properties": {
          "nodes": {
            "type": "array",
            "nullable": true
          }
        },
        "nullable": true
      },
      "CustomerModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "statusEnum": {
            "$ref": "#/components/schemas/CustomerStatusSample"
          },
          "concurrencyVersion": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CustomerStatusSample": {
        "enum": [
          "Unknown",
          "Active",
          "Inactive",
          "Suspended"
        ]
      },
      "CustomerUpdateStatusRequestModel": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "FilterCriteria": {
        "nullable": true
      },
      "FilterModel": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "noTracking": {
            "type": "boolean"
          },
          "orderings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilterOrderCriteria"
            },
            "nullable": true
          },
          "filters": {
            "type": "array",
            "nullable": true
          },
          "includes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "hierarchy": {
            "type": "string",
            "nullable": true
          },
          "hierarchyMaxDepth": {
            "type": "integer",
            "format": "int32"
          },
          "_FilterCriteriaSchema": {
            "$ref": "#/components/schemas/FilterCriteria"
          },
          "_CompositeSpecificationSchema": {
            "$ref": "#/components/schemas/CompositeSpecification"
          },
          "_SpecificationNodeSchema": {
            "$ref": "#/components/schemas/SpecificationNode"
          }
        }
      },
      "FilterOrderCriteria": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "nullable": true
          },
          "direction": {
            "$ref": "#/components/schemas/OrderDirection"
          }
        }
      },
...
      "OrderDirection": { },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ProblemError": {
        "type": "object",
        "additionalProperties": { },
        "description": "Represents an individual error entry with flexible structure for additional data"
      },
      "ResultProblemData": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ResultProblemResult"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemError"
            },
            "description": "Collection of additional errors outside the result object",
            "nullable": true
          }
        },
        "additionalProperties": { },
        "description": "Data section included within a problem details response",
        "nullable": true
      },
      "ResultProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "A URI reference that identifies the problem type",
            "nullable": true
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "description": "A human-readable explanation specific to this occurrence",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "description": "A URI reference that identifies the specific occurrence of the problem",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/ResultProblemData"
          }
        },
        "additionalProperties": { },
        "description": "A Problem Details response containing result information and contextual data (RFC 7807)"
      },
      "ResultProblemError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "nullable": true
          }
        },
        "additionalProperties": { },
        "description": "Represents an individual error entry with a message and optional metadata"
      },
      "ResultProblemResult": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean",
            "description": "Indicates whether the operation completed successfully"
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Collection of informational or diagnostic messages from the operation",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultProblemError"
            },
            "description": "Collection of domain or technical errors encountered during the operation",
            "nullable": true
          }
        },
        "additionalProperties": { },
        "description": "Represents the operation outcome with success status, messages, and errors",
        "nullable": true
      },
      "SpecificationNode": {
        "nullable": true
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "oauth2",
        "description": "JWT Bearer token authentication using OAuth 2.0"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}